Create  REGISTRATION SYSTEM
Description:

The Student Registration System is a comprehensive application designed to streamline the process of course registration for students within an educational institution. Developed as part of a Java Development internship , this system offers a user-friendly interface and robust functionality to meet the needs of both students and administrators. 

Check out our LinkedIn post here for a guide on how to run the code and explore its functionalities!   

Link   :    https://www.linkedin.com/posts/neha-kumari-b86292267_javadevelopment-internship-studentregistrationsystem-activity-7205414227824578560-Z_Sy?utm_source=share&utm_medium=member_android